Log even failed requests
It's not just the pingback action on the remote server, which can fail. Even the HTTP request itself may fail, and thus the $response would be an object of WP_Error, which currently causes fatal error.

In order to be able to log such a case to appropriate post's meta, we need to extract the `$post_id` from `$args['body']` earlier. Simply moving the `$args['body']` processing for `$pagelinkedfrom` and `$pagelinkedto` and then getting the post ID via `url_to_postid` would lead to unenecessarily execution of such a slow function (the `url_to_postid` can be slow).

Thus, this changeset moves the `url_to_postid` calls away from `$pagelinkedfrom`, but closer to `update_post_meta` calls, saving the `url_to_postid` call only to cases when it's needed.